Resolve duplicate game support models
